backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
sqlalchemy>=2.0.9
alembic>=1.10.3
geoalchemy2>=0.13.0
psycopg2-binary>=2.9.6
paho-mqtt>=1.6.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
websockets>=11.0.2
pytest>=7.3.1
httpx>=0.24.0
geojson>=3.0.1
shapely>=2.0.1 